💄 Edit: 러닝 거리 단위를 km로 변경

+++ RunUs/RunUs/Model/RunningInfo.swift
@@ -12,5 +12,5 @@ struct RunningInfo {
     var endDate: Date?
     var runningTime: String?
     var averagePace: String?
-    var distance: Int?
+    var distance: Double?
 }
